[FIX] evaluation: error handling on vectorize function
Given the formula =IF(A1:A20<>"", 1, 0) and A1:A20 is full of error values.

The result is currently a single #ERROR but it should be an array
of #ERROR

It happens because `errorHandlingCompute` is not part of `applyVectorization`.

As a result, when `applyVectorization` is manually called in IF, IFNA, etc.
Those use are not covered by the error handling code.

We fix this is master because it's a breaking change and it a lot code
moving around.

Task: 6009255
